+++ client/src/App.js
@@ -4,8 +4,8 @@ import { BrowserRouter, Routes, Route } from 'react-router-dom'
 import ScrollToTop from './util/ScrollToTop'
 import Header from "./components/Header";
 import Home from "../src/pages/Home";
-import Login from "../src/pages/Login";
-import SignUp from "../src/pages/SignUp";
+import Login from "../src/pages/Login,SignUp page/Login";
+import SignUp from "../src/pages/Login,SignUp page/SignUp";
 import Mypage from "../src/pages/Mypage";
 import Detail from "../src/pages/Detail";
 import Post from "../src/pages/Post";
